Extract client creation for reuse
